requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.10,<3.0.0
openai>=1.54.3
langchain-openai>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0
python-dotenv==1.0.0
pyyaml>=6.0
perplexity>=1.0.0
yfinance>=0.2.0
duckduckgo-search>=5.0.0

